fix MVC layering violations: remove model imports from GUI and simulation (#767)
- Move waveform defaults (DEFAULT_WAVEFORM_TYPE, default_waveform_params,
  format_waveform_spice_value) from simulation/waveform_utils.py to new
  models/waveform_defaults.py so models layer has zero simulation imports
- Update models/component.py to import waveform defaults from models layer
- Re-export from simulation/waveform_utils.py for backward compatibility
- Move GUI registration code (_register_graphics, _register_subcircuit_renderer,
  SubcircuitRenderer) from models/subcircuit_library.py to new
  GUI/subcircuit_gui_registration.py
- subcircuit_library.py now has zero imports from app/GUI/
- GUI/subcircuit_dialog.py calls register_subcircuit_gui() after model registration
- Update tests to import moved functions from GUI.subcircuit_gui_registration
